// fleetfinder.web.spa/src/app/models/enums/transport/cargo/cargo-body-kind.enum.ts
export enum CargoBodyKind {
  AutoTrain = 'Автопоезд',
  BeamVehicle = 'Балковоз',
  FlatbedVehicle = 'Бортовой',
  HydraulicLift = 'Гидроборт',
  CargoPassenger = 'Грузопассажирский',
  OversizedVehicle = 'Длинномер',
  IsothermalVehicle = 'Изотермический',
  ContainerVehicle = 'Контейнеровоз',
  FeedVehicle = 'Кормовоз',
  FlourVehicle = 'Муковоз',
  OpenVehicle = 'Открытый',
  PanelVehicle = 'Панелевоз',
  Pickup = 'Пикап',
  Pyramid = 'Пирамида',
  SemiTrailer = 'Полуприцеп',
  LightTrailer = 'Прицеп легковой',
  Refrigerator = 'Рефрижератор',
  RollVehicle = 'Рулоновоз',
  TractorUnit = 'Седельный тягач',
  AgriculturalGrainVehicle = 'Сельхозник/зерновоз',
  LivestockVehicle = 'Скотовоз',
  TimberVehicle = 'Сортиментовоз/лесовоз',
  GlassVehicle = 'Стекловоз',
  Coupling = 'Сцепка',
  TankContainer = 'Танк-контейнер',
  CurtainSider = 'Тентованный',
  InsulatedVehicle = 'Термобудка',
  PipeVehicle = 'Трубовоз',
  TrailerTruck = 'Фура',
  Van = 'Фургон',
  SolidMetal = 'Цельнометаллический',
  ChipCar = 'Щеповоз'
}