// fleetfinder.web.spa/src/app/models/enums/transport/cargo/cargo-transportation-kind.ts
export enum CargoTransportationKind {
  CargoTaxi = 'CargoTaxi',
  ApartmentMoving = 'ApartmentMoving',
  OfficeMoving = 'OfficeMoving',
  FurnitureTransport = 'FurnitureTransport',
  FoodTransport = 'FoodTransport',
  IntercityTransport = 'IntercityTransport',
  InternationalTransport = 'InternationalTransport',
  LCLTransport = 'LCLTransport',
  LivestockTransport = 'LivestockTransport',
  PersonalItemsTransport = 'PersonalItemsTransport',
  ApplianceTransport = 'ApplianceTransport',
  FruitTransport = 'FruitTransport',
  VegetableTransport = 'VegetableTransport',
  CountryHouseMoving = 'CountryHouseMoving',
  WarehouseMoving = 'WarehouseMoving',
  PianoTransport = 'PianoTransport',
  SafeTransport = 'SafeTransport',
  RefrigeratorTransport = 'RefrigeratorTransport',
  EquipmentTransport = 'EquipmentTransport',
  ConstructionMaterialsTransport = 'ConstructionMaterialsTransport',
  MotorcycleTransport = 'MotorcycleTransport',
}