// fleetfinder.web.spa/src/app/models/enums/transport/passenger/passenger-option.enum.ts
export enum PassengerOption {
  /// <summary>
  /// Аудио система и микрофон
  /// </summary>
  AudioSystemAndMicrophone = "AudioSystemAndMicrophone",

  /// <summary>
  /// Кондиционер
  /// </summary>
  AirConditioner = "AirConditioner",

  /// <summary>
  /// Система вентиляции
  /// </summary>
  VentilationSystem = "VentilationSystem",

  /// <summary>
  /// Большие багажные отделения
  /// </summary>
  LargeLuggageCompartments = "LargeLuggageCompartments",

  /// <summary>
  /// Монитор и DVD
  /// </summary>
  MonitorAndDVD = "MonitorAndDVD",

  /// <summary>
  /// Индивидуальное освещение
  /// </summary>
  IndividualLighting = "IndividualLighting",

  /// <summary>
  /// Подогрев сидений
  /// </summary>
  SeatHeating = "SeatHeating",

  /// <summary>
  /// Просторный салон
  /// </summary>
  SpaciousSalon = "SpaciousSalon",

  /// <summary>
  /// Ремни безопасности
  /// </summary>
  SafetyBelts = "SafetyBelts",

  /// <summary>
  /// Климат контроль
  /// </summary>
  ClimateControl = "ClimateControl",

  /// <summary>
  /// Туалет
  /// </summary>
  Toilet = "Toilet",

  /// <summary>
  /// Спальные места
  /// </summary>
  SleepingPlaces = "SleepingPlaces"
}