// fleetfinder.web.spa/src/app/models/enums/transport/passenger/passenger-facilities.enum.ts
export enum PassengerFacilities {
  /// <summary>
  /// Комфортабельный
  /// </summary>
  Comfortable= "Comfortable",

  /// <summary>
  /// Эконом класс
  /// </summary>
  Economy = "Economy",

  /// <summary>
  /// Стандартный
  /// </summary>
  Standard = "Standard"
}